change location of dump/ direcs
Edited to keep with the format of one 'dump' directory per instrument_type/year/ rather than how it was originally written in which each instrument_type/year/station/ had it's own 'dump'.  The potential advantage of the previous approach was that if people wanted the original monthly .csv files for the station they would be available but I found it unlikely that this would be a common request.  The disadvantage was that you'd be downloading a bunch of unnecessary files or I'd have to write a workaround later to exclude the dump folder when downloading.